+++ src/osu/scaling_factor.rs
@@ -4,6 +4,7 @@ use super::NORMALIZED_RADIUS;
 
 const OBJECT_RADIUS: f32 = 64.0;
 
+#[derive(Debug)]
 pub(crate) struct ScalingFactor {
     adjusted_factor: f32,
     factor: f32,